add friends projet, FriendRequest model
