+++ py_backend/friends/views.py
@@ -10,12 +10,16 @@ from .models import FriendRequest
 
 @require_http_methods(["POST"])
 def send_friend_request(request, user_id):
-    to_user = get_object_or_404(CustomUser, pk=user_id)
-    FriendRequest.objects.create(from_user=request.user, to_user=to_user, status='pending')
-    return JsonResponse({'status': 'success'}, status=200)
+	try:
+		to_user = get_object_or_404(CustomUser, pk=user_id)
+	except:
+		return JsonResponse(data={'errors': "Invalid JSON format"}, status=400)
+	FriendRequest.objects.create(from_user=request.user, to_user=to_user, status='pending')
+	return JsonResponse({'status': 'success'}, status=200)
 
 @require_http_methods(["POST"])
 def add_friend(self, friend):
     if not friend in self.friends.all():
         self.friends.add(friend)
         self.save()
+        return JsonResponse({'status': 'success'}, status=200)
