+++ py_backend/py_backend/settings.py
@@ -46,6 +46,7 @@ INSTALLED_APPS = [
 	'multiplayer',
 	'bootstrap5',
 	'users',
+	'friends',
 	# 'django_extensions',
 ]
