+++ py_backend/users/test/test_views.py
@@ -1,7 +1,7 @@
 from django.test import TestCase
 from django.urls import reverse
 from django.http import JsonResponse
-from users.models import CustomUser, FriendRequest
+from users.models import CustomUser
 from django.test import Client
 import json
 
@@ -503,21 +503,3 @@ class CSRFTokenTest(TestCase):
         self.assertEqual(response.status_code, 405)
 
 # =========================================================================================
-
-class FriendRequest(TestCase):
-    def setUp(self):
-        CustomUser.objects.create(
-            username="osterga",
-            email="osterga@gmail.com",
-            password="Damiendubocal75")
-        CustomUser.objects.create(
-            username="lboulatr",
-            email="lboulatr@gmail.com",
-            password="Damiendubocal75")
-    
-    def test_two_users_in_database(self):
-        user_count = CustomUser.objects.count()
-        self.assertEqual(user_count, 2)
-    
-    # def test_send_friend_request(self):
-
